requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
line-bot-sdk==3.11.0
google-generativeai==0.7.2
python-dotenv==1.0.1
